rag_pdf_project/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altgraph @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.3
asgiref==3.11.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==25.4.0
backcall==0.2.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bleach==6.2.0
build==1.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.1.8
coloredlogs==15.0.1
comm==0.2.0
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
docopt==0.6.2
durationpy==0.10
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.21.2
filelock==3.19.1
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
future @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.47.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.1
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib-metadata==6.8.0
importlib_resources==6.5.2
ipykernel==6.26.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab_pygments==0.3.0
kubernetes==35.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.83
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
langsmith==0.4.37
macholib @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.2.0
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.9.2
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.19.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pipreqs==0.5.0
platformdirs==4.0.0
posthog==5.4.0
prompt-toolkit==3.0.41
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==5.9.6
psycopg==3.2.13
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
Pygments==2.17.2
PyMuPDF==1.26.5
pyparsing==3.3.1
PyPika==0.50.0
pyproject_hooks==1.2.0
python-dateutil==2.8.2
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==25.1.1
referencing==0.36.2
regex==2026.1.14
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
shellingham==1.5.4
six @ file:///AppleInternal/Library/BuildRoots/2c89a47b-9dd5-11ef-938f-6e654a286000/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
soupsieve==2.8.3
SQLAlchemy==2.0.45
sqlparse==0.5.5
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.2.2
tornado==6.3.3
tqdm==4.67.1
traitlets==5.13.0
transformers==4.57.6
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.12
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
yarg==0.1.9
yarl==1.22.0
zipp==3.17.0
zstandard==0.25.0
